KVDB - Review feedback. Fixing tests
