+++ src/engine/test/engineTestMain.cpp
@@ -1,0 +1,16 @@
+#include <gtest/gtest.h>
+
+#include <kvdb/kvdbManager.hpp>
+#include <logging/logging.hpp>
+
+int main(int argc, char** argv)
+{
+    KVDBManager::init("/tmp/");
+
+    logging::LoggingConfig logConfig;
+    logConfig.logLevel = logging::LogLevel::Off;
+    logging::loggingInit(logConfig);
+
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
