+++ src/engine/test/source/kvdb/CMakeLists.txt
@@ -2,6 +2,6 @@ add_executable(kvdb_test
   ${TEST_SOURCE_DIR}/kvdb/kvdb_test.cpp
 )
 
-target_link_libraries(kvdb_test kvdb gtest base)
+target_link_libraries(kvdb_test kvdb engine_test_main base)
 
 gtest_discover_tests(kvdb_test)
