+++ src/engine/test/source/kvdb/kvdb_test.cpp
@@ -389,15 +389,3 @@ TEST_F(KVDBTest, KVDBConcurrency)
     KVDBManager::get().deleteDB(dbName);
 }
 } // namespace
-
-int main(int argc, char** argv)
-{
-    KVDBManager::init("/tmp/");
-
-    logging::LoggingConfig logConfig;
-    logConfig.logLevel = logging::LogLevel::Off;
-    logging::loggingInit(logConfig);
-
-    ::testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
-}
